+++ CHANGELOG.md
@@ -7,12 +7,14 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 ### Enhancements
 
 - Localizes all alerts into 41 languages.
+- Makes sure to refresh free trial eligibility on every paywall open.
 
 ### Fixes
 
 - Makes `device.isSandbox` more reliable.
 - Fixes the web restore alert not showing the "Yes" action button and "Cancel" incorrectly triggering the restore action.
 - Fixes a rare issue where a user's subscription could remain active after a refund, preventing paywalls from being shown.
+- Fixed trial eligibility for Stripe products.
 
 ## 4.14.0